dv project/requirement.txt
streamlit
pandas
numpy
matplotlib.pyplot
seaborn
plotly
scikit-learn
